Reorganize router utils
