+++ packages/router/src/active-route-loader.tsx
@@ -2,13 +2,14 @@ import React, { useRef, useState, useEffect } from 'react'
 
 import { unstable_batchedUpdates } from 'react-dom'
 
+import { getAnnouncement, getFocus, resetFocus } from './a11yUtils'
 import {
   ActivePageContextProvider,
   LoadingStateRecord,
 } from './ActivePageContext'
 import { PageLoadingContextProvider } from './PageLoadingContext'
 import { useIsMounted } from './useIsMounted'
-import { Spec, getAnnouncement, getFocus, resetFocus } from './util'
+import { Spec } from './util'
 
 import { ParamsProvider, useLocation } from '.'
